Add Burgers' equation test
Add Burgers' equation test. This is more complex test with respect
the Lorenz' one.

The test seems to work with the forward explicit Euler integrator,
but not with the RK class. I do not concern about this beacuse the
FOODiE is not yet stable thus the RK class of integrators are not
yet stable too. Just rember there is bug in RK class or into the
Burgers test when coupled with RK class integrators.