Unsuccessful fix of overloaded-operartors issue
Intel fortran (up to 15.x version) shows an issue with operators
overloading, e.g. `field = field + dt*field%t()` does not compile
while `field = field + field%t()*dt` does.

A problem was due to the overloading of also type names with constructor
functions and this hotfix is based on eliminating those overloads, but
it is not enough.

Why:

It is not possible with Intel 15.x branch to compile a complete
overloaded field.

This change addresses the need by:

Removing overloading of derived types with constructor functions in
flavor of `init` type bound procedures.

Side effects:

Update the burgers, lorenz and oscillation tests.