Fix issue#18
Fix memory management of previous time steps for the three test cases,
Burgers, Lorenz and Oscillation. As a matter of fact, the oscillation
test shows wrong results whit Adams-Bashforth class.

Why:

Multi-step methods, namely the Adams-Bashforth being the only currently
implemented, show bad results for oscillation test. The problem rely on
the writing of overloaded-operators-results into the current state
solution while the integrator is using it, thus with a wrong summation
of *previous tendencies*.

This change addresses the need by:

Nothing is changed into the integrators implementation: the
Adams-Bashforth is unchanged, thus it is probably well implemented. The
patch is applied only on the *client codes*, namely the tests (Burgers,
Lorenz and Oscillation): the previous time steps solution are stored
into a different array-member of the extended `type_integrand` object
thus:

+ overloaded-operators procedures operates **only** on the `state` array
member;
+ time-derivative and previous-time-steps procedures are the only public
methods allowed to operate on also the `previous` array member.

Side effects:

The tests sources must be changed: the `previous` array has been added
as a member of the field. It is worth noting that for avoiding wrong,
concurrent writing on the up-to-date state array when
overloaded-operators are used within the time-derivative procedure, the
up-to-date state array is also copied into the upper-bound of `previous`
array, thus there is a not efficient use of memory that should be
investigated during the performance analysis.