Add WenOOF external dependency
Add WenOOF external dependency

Why:

For implementing high order integration on Euler 1D test.

This change addresses the need by:

Clone the WenOOF project into the external. It is not possible to add it
as *git submodule* because the two project are related: FOODiE will be
necessary for the WenOOF tests suite.

Side effects:

Git pull will be necessary for updating WenOOF external.