Add time argument to residuals function
Modify the API of the time derivative (residuals function) abstract
interface in order to allow the passage of also the time.

Why:

Previously the abstract *time derivative* function has as dummy argument
only the (optional) time level *step*, that is used for example into
multi-step schemes. However, there are cases where the residuals
function directly depends on also the time and not only on the integrand
variable(s), i.e. in general `R(t,U(t))`. For this reason the API of the
time derivative function has been improved (generalized) allowing the
(optional) passage of also the time.

This change addresses the need by:

Modify the abstract interface of the time derivative function adding the
(optional) dummy argument *t* (time).

Side effects:

Modify all the solvers and tests accordingly with the new API.