Add OpenMP 1D Euler test
Add a first test for parallel benchmark with OpenMP directives.
It is based on the 1D Euler test, it being the most complex test
currently implemented.

Why:

It is crucial establish the parallel performance of FODDiE and
address the eventual drawback of the Abstract Calculus Pattern*

This change addresses the need by:

Add the euler-1D-openmp test.

Side effects:

Remove the *pure* attributes to the type_integrand procedures.