Add Euler 1D solver app
Add Euler 1D solver app: it is a complete solver for the 1D Euler PDEs
system (with multi-species capability).

Why:

The app is usefull to develop new integrator/reconstructor for
hyperbolic PDEs systems and it is an optimal test for the library.

Side effects:

All submodules have updated, but the tests suite has not be checked...